requirements.txt
# Web Framework
fastapi==0.109.0
uvicorn[standard]==0.27.0
pydantic==2.5.3
pydantic-settings==2.1.0

# Database
sqlalchemy==2.0.25
alembic==1.13.1

# Task Queue
celery==5.3.6
redis==5.0.1

# Web Scraping
beautifulsoup4==4.12.3
requests==2.31.0
lxml==5.1.0
scrapy==2.11.0
selenium==4.17.2
playwright==1.41.0

# PDF Processing
PyPDF2==3.0.1
pdfplumber==0.10.3
pypdf==4.0.1
Pillow==10.2.0
img2pdf==0.5.1

# Etsy API
etsy-python==0.3.0

# Configuration
pyyaml==6.0.1
python-dotenv==1.0.0

# Utilities
aiofiles==23.2.1
httpx==0.26.0
tenacity==8.2.3

# Testing (optional)
pytest==7.4.4
pytest-asyncio==0.23.3
pytest-cov==4.1.0

# Docker
docker==7.0.0
